sweagent/requirements.txt
# Core dependencies
anthropic>=0.7.0
langchain-anthropic>=0.1.0
langchain-core>=0.1.0
langgraph>=0.1.0
langgraph-checkpoint>=1.0.0
pydantic>=2.0.0

# File and data processing
pyyaml>=6.0
requests>=2.28.0

# Optional dependencies
# Uncomment if needed
# jupyter>=1.0.0           # For notebook support  
# matplotlib>=3.5.0        # For plotting and visualization
# pandas>=1.5.0            # For data analysis tasks
# numpy>=1.21.0            # For numerical computations

# Development dependencies (optional)
# pytest>=7.0.0            # For testing
# black>=22.0.0            # For code formatting
# flake8>=5.0.0            # For linting
# mypy>=1.0.0              # For type checking